+++ cmd/web/renders.go
@@ -87,6 +87,7 @@ func (app *application) addDefaultData(td *templateData, req *http.Request) *tem
 	td.API["StatsTotals"] = APIStatsTotals
 	td.API["StatsTotalsByType"] = APIStatsTotalsByType
 	td.API["StatsTotalsByClass"] = APIStatsTotalsByClass
+	td.API["StatsLimits"] = APIStatsLimits
 	td.API["Map"] = APIMap
 	td.API["MapData"] = APIMapData
 	td.API["Licensing"] = APILicensing